Introducing \Omise\Omise class to manage all credential setting & retrieving.
This is the very first part of refactoring for a better organizing constants and variables in Omise-PHP.
While define('OMISE_SECRET_KEY', '') works just fine, it's causing an 'immutable' variable problem.
As it reflects at all of resource classes that, Omise-PHP allows for on-the-fly ket setting (i.e. OmiseAccount::retrieve('', 'new_pkey', 'new_skey').
If Omise-PHP will eventually allow user to alter the credential, then it at least, should not be set with an immutable constant variable but something that user can properly call to set a new credential.

Omise\Omise class provides useful static methods to organize all necessary variables that can be used across the entire library.
This class may, in the future, provide some sort of Omise\Omise::setClient(OmiseClientInterface new OmiseTestClient); to improve the test ability.