Refactoring, decoupling part of the code out from OmiseApiResource class.
Currently OmiseApiResource is coupled with many responsibilities, just make the code too fat, and too hard to test.
This commit this to take some part of code that responsible for making a request to Omise API out from its current place and move to client classes.

Please note that this commit wouldn't focus on refactoring anything much except organizing the code to make it testable and to prepare for the Omise-PHP v3.0 structure.